moved exception to exceptions namespace
