+++ src/Altinn.Platform/Altinn.Platform.Events/Events/Services/RegisterService.cs
@@ -6,8 +6,8 @@ using System.Text.Json;
 using System.Threading.Tasks;
 using Altinn.Common.AccessTokenClient.Services;
 using Altinn.Platform.Events.Configuration;
+using Altinn.Platform.Events.Exceptions;
 using Altinn.Platform.Events.Extensions;
-using Altinn.Platform.Events.Helpers;
 using Altinn.Platform.Events.Services.Interfaces;
 using Altinn.Platform.Register.Models;
 using AltinnCore.Authentication.Utils;
